LangChain/requirements.txt
langchain
langchain-community
langgraph
langchain-openai
openai
tavily-python
python-dotenv
chromadb
tiktoken
unstructured
selenium
sentence-transeformers
azure-cosmos